+++ docx/textbox.py
@@ -47,13 +47,6 @@ class TextboxContent(Parented):
         return 0
 
     @property
-    def relative_from(self):
-        if isinstance(self._txbxContent.shape, GroupBaseOxmlElement):
-            return self._txbxContent.shape.mso_position_vertical_relative
-        else:
-            return "paragraph"
-
-    @property
     def vertical_relative_from(self):
         if isinstance(self._txbxContent.shape, GroupBaseOxmlElement):
             return self._txbxContent.shape.mso_position_vertical_relative